feat(DATAGO-114372): Added SAM A2A proxy to connect to remote A2A agents (#392)
This PR adds a new component type to SAM: Proxy

The Proxy component is used to bring 3rd party agents into SAM. In this PR we introduce a single proxy type: A2A agents

The A2A proxy will allow SAM to be configured to connect to an external A2A agent over HTTPS and interact with it. It has a few different authentication types (api key, OAuth client credentials and bearer token). It also will allow for artifact loading and saving so that external agents can be given artifacts from the artifact store.