chore(DATAGO-114919): Fix file uploads for files larger than 1MB (#413)
Files that require using the artifacts endpoint to upload files were broken. This PR addresses that and fixes the tests

It also became apparent during the fix that audio transcription was returning large output directly to the LLM. That is bad and not intentional. That fix is in here too